% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquavalidator-package.R
\docType{package}
\name{aquavalidator-package}
\alias{aquavalidator}
\alias{aquavalidator-package}
\title{aquavalidator: AquaValidator}
\description{
A GUI to clearly present water analysis results and provide tools to help judge whether these results are accurate and realistic.
}
\author{
\strong{Maintainer}: Koen van der Heide \email{kheide@aqualysis.nl}

}
\keyword{internal}
