% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startValidator}
\alias{startValidator}
\title{Start Aqualysis Validator met extra opties}
\usage{
startValidator(browser = TRUE, ...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Start Aqualysis Validator met extra opties
}
